Safari-609.1.11.1/LayoutTests/http/tests/xmlhttprequest/access-control-sandboxed-iframe-denied-without-wildcard-expected.txt

CONSOLE MESSAGE: line 16: Origin null is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 16: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-sandboxed-iframe-denied-without-wildcard.cgi due to access control checks.
This test verifies that sandboxed iframe does not have XmlHttpRequest access to its server with "Access-Control-Allow-Origin" set to its own origin (127.0.0.1).

This test will print "PASS" on success.



--------
Frame: '<!--frame1-->'
--------
PASS: Exception thrown. Sandboxed iframe XHR access was denied in 'send'. [ A network error occurred.].
